# Conflicts:
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/common/adapter/scherec/convert/ConvertApplicationToShare.java
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/common/service/other/CollectAchievementImpl.java
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/stamp/AppCommonDomainServiceImp.java
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.infra/src/main/java/nts/uk/ctx/at/request/infra/entity/setting/company/applicationapprovalsetting/stampsetting/KrqmtAppStamp.java
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.infra/src/main/java/nts/uk/ctx/at/request/infra/repository/application/stamp/JpaAppStampSettingRepository.java
#	nts.uk/uk.at/at.file/nts.uk.file.at.app/src/main/java/nts/uk/file/at/app/export/worktype/PreparationBeforeApplyExportImpl.java
#	nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/view/kaf/002/a/kaf002.a.vm.ts
#	nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/view/kaf/002/c/kaf002.c.vm.ts
#	nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/view/kaf/002/m/kaf002.m.vm.ts
#	nts.uk/uk.com/shr/nts.uk.shr.com/src/main/java/nts/uk/shr/com/program/ProgramIdConsts.java
#	nts.uk/uk.com/shr/nts.uk.shr.com/src/main/java/nts/uk/shr/com/program/ProgramsManager.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/byperiod/KrcdtAnpAttendanceTime.java
@@ -136,11 +136,11 @@ public class KrcdtAnpAttendanceTime extends ContractUkJpaEntity implements Seria
 
 	/** 就業時間金額 */
 	@Column(name = "WORK_TIME_AMOUNT")
-	public int workTimeAmount;
+	public long workTimeAmount;
 	
 	/** 割増金額合計 */
 	@Column(name = "PREMIUM_AMOUNT_TOTAL")
-	public int premiumAmountTotal;
+	public long premiumAmountTotal;
 	
 	/** 就業時間 */
 	@Column(name = "WORK_TIME")
